requirements.txt
appdirs==1.4.3
Django==1.8
djangorestframework==3.4.0
Mako==1.0.6
MarkupSafe==1.0
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
six==1.10.0
